BaseEntity --> com.degloba.domain.support.BaseEntity Entity  --> com.degloba.domain.seedwork.Entity Serializable  --> Long Integer  --> Long MyEntity extends AbstractEntity   -->  MyEntity extends BaseAggregateRoot IRepository --> EntityRepository

+++ trunk/degloba-domain/src/test/java/com/degloba/domain/AbstractEntityTest.java
@@ -45,7 +45,8 @@ public class AbstractEntityTest {
 
     @Test
     public void testVersionAccessor() {
-        int version = 10;
+        //int version = 10;
+    	Long version = (long) 10;
         instance.setVersion(version);
         assertEquals(version, instance.getVersion());
     }
